Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/requestidlecallback/callback-exception-expected.txt

CONSOLE MESSAGE: line 19: Error: requestIdleCallbackException

PASS requestIdleCallback callback exceptions are reported to error handler 
